backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
feedparser==6.0.11
beautifulsoup4==4.12.2
sgmllib3k==1.0.0
chatterbox-tts
torch
torchaudio
